✨ 마일 스톤 페이지 최종 구현 완료.
