+++ fe/src/api/issues.js
@@ -1,3 +1,3 @@
 import { BASE_URL } from '.';
 
-export const ISSUES_URL = `${BASE_URL}/`;
+export const ISSUES_URL = `${BASE_URL}/home`;
